voice/requirements.txt
# SION Voice Cloning Module Dependencies
# 교차언어 음성 클로닝을 위한 의존성

# Core ML/DL
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0
numpy>=1.24.0
pydub>=0.25.0
pyworld>=0.3.4

# Speech Processing
g2pk>=0.9.4                 # Korean grapheme-to-phoneme
jaconv>=0.3.4               # Japanese character conversion
fugashi>=1.3.0              # Japanese morphological analysis
unidic-lite>=1.0.8          # Japanese dictionary

# GPT-SoVITS specific
cn2an>=0.5.22               # Chinese number conversion
pypinyin>=0.49.0            # Chinese pinyin
jieba>=0.42.1               # Chinese segmentation
LangSegment>=0.2.0          # Language segmentation
wordsegment>=1.3.1

# Audio playback
pygame>=2.5.0
sounddevice>=0.4.6

# API & Utils
fastapi>=0.104.0
uvicorn>=0.24.0
pyyaml>=6.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Hugging Face Hub
huggingface_hub>=0.19.0

# Optional: For training
# tensorboard>=2.14.0
# wandb>=0.16.0

